Added optparse-applicative commandline.
Still to be done here is exporting an API for extension.  We should be
able to build programs that can handle their own arguments then hand off
to diagrams `defaultMain`.

Also missing is version information.  We will need to look at how
Cmdargs does this.

Once we figure this out we can extract all of this into a common place
shared by all the backends.