Split options record into records for specific main's.
Optparse-Applicative allows us to combine records and parsers at each
form of main that we offer.  The record doesn't need to be monolithic
but we can just include options for the specific needs of a particular
main.

One issue now is that normally "-o" is not optional, but we want to
allow in the `multiMain` a lone "--list".