Implement typeclasses for mainWith.
The idea is to allow users to create programs that generate diagrams
based on some parameters taken on the command line.  This also allows
for a single `mainWith` that can handle a single diagram, multiple named
diagrams, diagrams for each page, or animations with numbered output
files.  Arguments can be taken in order from the commandline or with a
custom parser from the command line or with a custom named argument via
the `Parseable` typeclass.

All this is experimental in this branch and I'm working to factor out
the code that will be common to all backends (or useful in general) and
put it in -lib.